+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 173:
 
 * Remove Autobahn testsuite doc note
 * Fix buffers_adapter iterator value type
+* Fix buffers_adapter max_size
 
 --------------------------------------------------------------------------------